...by exactly +1 or -1 to pull the moons together.
For example, if Ganymede has an x position of 3, and Callisto has a
x position of 5, then Ganymede's x velocity changes by +1 (because
5 > 3) and Callisto's x velocity changes by -1 (because 3 < 5).
However, if the positions on a given axis are the same, the velocity
on that axis does not change for that pair of moons.

A LOT going on here.